+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.0.2
 
 - Correct pipeline names
+- Correct container prefixes
 
 ## 3.0.1